+++ components/clarinet-cli/examples/cbtc/tests/cbtc-token_test.ts
@@ -1,15 +1,32 @@
-
-import { Clarinet, Tx, Chain, Account, types } from 'https://deno.land/x/clarinet@v1.0.2/index.ts';
-import { assertEquals } from 'https://deno.land/std@0.90.0/testing/asserts.ts';
+import {
+  Clarinet,
+  Tx,
+  Chain,
+  types,
+} from "https://deno.land/x/clarinet@v1.5.4/index.ts";
+import { assertEquals } from "https://deno.land/std@0.191.0/testing/asserts.ts";
 
 Clarinet.test({
-    name: "Ensure that tokens can be minted and burnt",
-    fn(chain: Chain, accounts: Map<string, Account>) {
-        let block = chain.mineBlock([
-            Tx.contractCall("cbtc-token", "mint", [types.uint(1000), types.principal("ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG")], "ST1SJ3DTE5DN7X54YDH5D64R3BCB6A2AG2ZQ8YPD5"),
-            Tx.contractCall("cbtc-token", "burn", [types.uint(500)], "ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG")
-        ]);
-        assertEquals(block.receipts.length, 2);
-        assertEquals(block.height, 3);
-    },
+  name: "Ensure that tokens can be minted and burnt",
+  fn(chain: Chain) {
+    const block = chain.mineBlock([
+      Tx.contractCall(
+        "cbtc-token",
+        "mint",
+        [
+          types.uint(1000),
+          types.principal("ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG"),
+        ],
+        "ST1SJ3DTE5DN7X54YDH5D64R3BCB6A2AG2ZQ8YPD5"
+      ),
+      Tx.contractCall(
+        "cbtc-token",
+        "burn",
+        [types.uint(500)],
+        "ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG"
+      ),
+    ]);
+    assertEquals(block.receipts.length, 2);
+    assertEquals(block.height, 3);
+  },
 });